requirements.txt
Django
psycopg2
django-search-views
